+++ common/src/lib.rs
@@ -3,4 +3,5 @@ pub use time;
 pub mod dc09;
 pub mod logging;
 pub mod scenarios;
+pub mod tls;
 pub mod utils;